+++ src/main/java/com/dataracy/user/infra/jwt/JwtFilter.java
@@ -3,8 +3,8 @@ package com.dataracy.user.infra.jwt;
 import com.dataracy.common.exception.BusinessException;
 import com.dataracy.common.util.ExtractHeaderUtil;
 import com.dataracy.user.domain.enums.RoleStatusType;
-import com.dataracy.user.infra.auth.CustomUserDetails;
-import com.dataracy.user.infra.auth.UserAuthentication;
+import com.dataracy.user.infra.security.CustomUserDetails;
+import com.dataracy.user.infra.security.UserAuthentication;
 import com.dataracy.user.status.AuthErrorStatus;
 import com.dataracy.user.status.AuthException;
 import jakarta.servlet.FilterChain;
