+++ src/main/java/com/dataracy/user/infra/security/CustomUserDetails.java
@@ -1,4 +1,4 @@
-package com.dataracy.user.infra.auth;
+package com.dataracy.user.infra.security;
 
 import com.dataracy.user.domain.enums.RoleStatusType;
 import lombok.Getter;
